requirements.txt
streamlit
pandas
opencv-python
numpy
Pillow
